+++ crates/opsml_sql/src/enums/client.rs
@@ -107,7 +107,7 @@ impl SqlClient for SqlClientEnum {
     async fn query_page(
         &self,
         sort_by: &str,
-        page: i64,
+        page: i32,
         search_term: Option<&str>,
         repository: Option<&str>,
         table: &CardSQLTableNames,
