+++ crates/opsml_sql/src/sqlite/client.rs
@@ -730,7 +730,7 @@ impl SqlClient for SqliteClient {
     async fn query_page(
         &self,
         sort_by: &str,
-        page: i64,
+        page: i32,
         search_term: Option<&str>,
         repository: Option<&str>,
         table: &CardSQLTableNames,
